migreate deploy function to main page
