+++ client/components/+deploy/deploy-details.scss
