+++ client/components/+deploy/deploy.scss
@@ -1,0 +1,10 @@
+.Deploys {
+    .TableCell {
+      &.name {
+        flex-grow: 1.7;
+      }
+      &.namespace{
+        flex-grow: 0.3;
+      }
+    }
+  }
