+++ client/components/+deploy/deploy.tsx
@@ -1,20 +1,78 @@
-import * as React from "react";
+import "./deploy.store.ts";
+
+import React from "react";
 import { observer } from "mobx-react";
-import { Redirect, Route, Switch } from "react-router";
-import { RouteComponentProps } from "react-router-dom";
+import { RouteComponentProps } from "react-router";
 import { Trans } from "@lingui/macro";
+import { Deploy, deployApi } from "../../api/endpoints";
+import { KubeObjectMenu, KubeObjectMenuProps } from "../kube-object/kube-object-menu";
 import { MainLayout, TabRoute } from "../layout/main-layout";
-interface Props extends RouteComponentProps {
-} 
+import { KubeObjectListLayout } from "../kube-object";
+import { IDeployWorkloadsTemplateParams } from "../+deploy";
+import { apiManager } from "../../api/api-manager";
+import { deployStore } from "./deploy.store";
+
+enum sortBy {
+    name = "name",
+    namespace = "namespace",
+    generateTimestamp = "generateTimestamp",
+    age = "age",
+}
+
+interface Props extends RouteComponentProps<IDeployWorkloadsTemplateParams> {
+}
 
 @observer
-export class Deploy extends React.Component<Props>{
-    
-    render(){
+export class Deploys extends React.Component<Props> {
+
+    render() {
         return (
             <MainLayout>
-                <div>todo...</div>
-           </MainLayout>
+                <KubeObjectListLayout
+                    className="Deploys" store={deployStore}
+                    sortingCallbacks={{
+                        [sortBy.name]: (deploy: Deploy) => deploy.getName(),
+                        [sortBy.namespace]: (deploy: Deploy) => deploy.getNs(),
+                        [sortBy.age]: (deploy: Deploy) => deploy.getAge(false),
+                    }
+                    }
+
+                    searchFilters={
+                        [
+                            (deploy: Deploy) => deploy.getSearchFields(),
+                        ]}
+
+                    renderHeaderTitle={< Trans > Deploys </Trans >}
+                    renderTableHeader={
+                        [
+                            { title: <Trans>Name</Trans>, className: "name", sortBy: sortBy.name },
+                            { title: <Trans>Namespace</Trans>, className: "namespace", sortBy: sortBy.namespace },
+                            { title: <Trans>GenerateTimestamp</Trans>, className: "generateTimestamp", sortBy: sortBy.generateTimestamp },
+                            { title: <Trans>Age</Trans>, className: "age", sortBy: sortBy.age },
+                        ]}
+
+                    renderTableContents={(deploy: Deploy) => [
+                        deploy.getName(),
+                        deploy.getNs(),
+                        new Date(deploy.getGenerateTimestamp() * 1e3).toISOString(),
+                        deploy.getAge(),
+                    ]}
+
+                    renderItemMenu={(item: Deploy) => {
+                        return <DeployMenu object={item} />
+                    }}
+                />
+            </MainLayout>
         )
     }
 }
+
+export function DeployMenu(props: KubeObjectMenuProps<Deploy>) {
+    return (
+        <KubeObjectMenu {...props} />
+    )
+}
+
+apiManager.registerViews(deployApi, {
+    Menu: DeployMenu,
+})
