--- client/components/+workloads-deploy/deploy.scss
@@ -1,10 +1,0 @@
-.Injectors {
-  .TableCell {
-    &.name {
-      flex-grow: 1.7;
-    }
-    &.namespace{
-      flex-grow: 0.3;
-    }
-  }
-}
