--- client/components/+workloads-deploy/index.ts
@@ -1,2 +1,0 @@
-export * from "./deploy"
-export * from "./deploy-details"
