+++ client/components/app.tsx
@@ -16,7 +16,7 @@ import { usersManagementRoute } from "./+user-management/user-management.routes"
 import { clusterRoute, clusterURL } from "./+cluster";
 import { KubeConfigDialog } from "./kubeconfig-dialog/kubeconfig-dialog";
 import { Nodes, nodesRoute } from "./+nodes";
-import { Deploy, deployURL, deployRoute } from "./+deploy";
+import { Deploys, deployURL, deployRoute } from "./+deploy";
 import { Workloads, workloadsRoute, workloadsURL } from "./+workloads";
 import { Tenant, tenantURL,tenantRoute } from "./+tenant";
 import { Namespaces, namespacesRoute } from "./+namespaces";
@@ -64,7 +64,7 @@ class App extends React.Component {
                     <Route component={Cluster} {...clusterRoute}/>
                     <Route component={Nodes} {...nodesRoute}/>
                     <Route component={DeploymentEngine} {...deploymentEngineRoute} />
-                    <Route component={Deploy} {...deployRoute}/>
+                    <Route component={Deploys} {...deployRoute}/>
                     <Route component={Workloads} {...workloadsRoute}/>
                     <Route component={Config} {...configRoute}/>
                     <Route component={Network} {...networkRoute}/>
